% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePI.R
\name{computePI}
\alias{computePI}
\title{ComputePI}
\usage{
computePI(obsdata = NULL, simdata, stratify = NULL, TIME = TIME,
  DV = DV, REPL = REP, LLOQ = NULL, NBINS = NULL,
  bin_by_strata = TRUE, style = "ntile", breaks = NULL,
  cut_right = FALSE, filterblq = FALSE, predcorrection = FALSE,
  predcorrection_islogdv = FALSE, predcorrection_lowerbnd = 0,
  PI = c(0.05, 0.5, 0.95), CI = c(0.025, 0.5, 0.975),
  bootstrapobsdata = FALSE)
}
\arguments{
\item{obsdata}{Obs dataset}

\item{simdata}{Sim dataset}

\item{stratify}{Stratify formula: right hand side only}

\item{TIME}{Time column: a name or an expression}

\item{DV}{Dependent Variable column: a name or an expression}

\item{REPL}{Replication Index column: a name}

\item{LLOQ}{LLOQ column: a name, an expression, or a fixed value}

\item{NBINS}{NBINS column: a name, an expression, or a fixed value}

\item{bin_by_strata}{Flag to indicate if the binning is by strata (default) or global}

\item{style}{Style of the binning procedure: "ntile" (default), {"fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", or "jenks"} (ClassInterval styles), "breaks" (user defined breaks)}

\item{breaks}{User defined breaks: either a vector or a wide data.frame with a column "breaks", and other colums corresponding to the stratify variables}

\item{cut_right}{Flag to indicate if the intervals should be closed on the right (and open on the left) or vice versa. Default to False, unlike the cut function.}

\item{filterblq}{Flag to remove the BLQ values just before computing the quantiles, but after "merging" with rep(, sim)}

\item{predcorrection}{Flag to indicate if the VPC has to be pred corrected or not}

\item{predcorrection_islogdv}{Flag to indicate if the data was log transformed}

\item{predcorrection_lowerbnd}{Not used yet}

\item{PI}{Prediction Intervals: vector of any number of  values}

\item{CI}{Confidence Intervals around the Prediction Intervals : vector of any number of value}

\item{bootstrapobsdata}{Not used yet}
}
\description{
ComputePI
}
