% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats2.R
\name{generics}
\alias{generics}
\alias{observed}
\alias{simulated}
\alias{censoring}
\alias{stratify}
\alias{binning}
\alias{predcorrect}
\alias{nopredcorrect}
\alias{vpcstats}
\title{Perform a Visual Predictive Check (VPC) computation}
\usage{
observed(o, ...)

simulated(o, ...)

censoring(o, ...)

stratify(o, ...)

binning(o, ...)

predcorrect(o, ...)

nopredcorrect(o, ...)

vpcstats(o, ...)
}
\arguments{
\item{o}{An object.}

\item{...}{Additional arguments.}
}
\description{
These functions work together to claculate the statistics that are plotted
in a VPC. They would typically be chained together using the "pipe" operator
(see Examples).
}
\examples{

\dontrun{
library(data.table)
library(magrittr)
library(vpc)

exampleobs <- as.data.table(vpc::simple_data$obs)[MDV == 0]
examplesim <- as.data.table(vpc::simple_data$sim)[MDV == 0]
exampleobs$PRED <- examplesim[REP == 1, PRED]
exampleobs$SEX <- rep(c("F", "M"), len=nrow(exampleobs))

vpc <- observed(exampleobs, x=TIME, y=DV) \%>\%
    simulated(examplesim, y=DV) \%>\%
    stratify(~ ISM) \%>\%
    binning(TIME) \%>\%
    predcorrect(pred=PRED) \%>\%
    vpcstats()

plot(vpc)

# Example with 2-way stratification

vpc <- vpc \%>\%
    stratify(SEX ~ ISM) \%>\%
    binning(TIME) \%>\%
    predcorrect(pred=PRED) \%>\%
    vpcstats()

plot(vpc)
}

}
