% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsdata.R
\docType{data}
\name{obsdata}
\alias{obsdata}
\title{observed Pharmacometrics PK/PD Data}
\format{A data frame with 7154 rows and 4 variables
\describe{
  \item{ID}{Subject Identifier, an integer from 57 to 186}
  \item{Time}{Time of sample measurement, in hours}
  \item{DV}{Dependent variable value}
  \item{DVTYPE}{type of Dependent variable with values "CObs","CMObs" or "EObs"}
}}
\source{
ASCPT Challenge competition
}
\usage{
obsdata
}
\description{
A dataset containing concentration-time data with the
type of dependent variable.
}
\keyword{datasets}
