% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats2.R
\name{print.vpcstatsobj}
\alias{print.vpcstatsobj}
\title{Print a \code{vpcstatsobj}.}
\usage{
\method{print}{vpcstatsobj}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments can be specified but are ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print a \code{vpcstatsobj}.
}
