% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{observed Pharmacometrics PK/PD Data}
\format{A data frame with 600 rows and 10 variables
\describe{
  \item{ID}{Subject Identifier, an integer from 57 to 186}
  \item{Time}{Time of sample measurement, in hours}
  \item{DV}{Dependent variable value}
  \item{DVTYPE}{type of Dependent variable with values "CObs","CMObs" or "EObs"}
  \item{REPL}{Simulation Repliate number, an integer from 0 to 50}
}}
\source{
ASCPT Challenge competition
}
\usage{
simdata
}
\description{
A dataset containing concentration-time data with the
given dose and some subject characteristics to help in the app exploration.
}
\keyword{datasets}
