% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{vpcstats.data.frame}
\alias{vpcstats.data.frame}
\title{vpcstats}
\usage{
\method{vpcstats}{data.frame}(obsdata = NULL, simdata, stratify = NULL,
  ID = ID, TIME = TIME, DV = DV, REPL = REP, PRED = PRED,
  LLOQ = NULL, NBINS = NULL, bin_by_strata = TRUE,
  bin_style = "ntile", breaks = NULL, cut_right = FALSE,
  quantile_type = 7, filterblq = FALSE, predcorrection = FALSE,
  predcorrection_islogdv = FALSE, predcorrection_lowerbnd = 0,
  PI = c(0.05, 0.5, 0.95), CI = c(0.025, 0.5, 0.975),
  bootstrapobsdata = FALSE)
}
\arguments{
\item{obsdata}{Obs dataset}

\item{simdata}{Sim dataset}

\item{stratify}{Stratify formula: right hand side only}

\item{TIME}{Time column: a name or an expression}

\item{DV}{Dependent Variable column: a name or an expression}

\item{REPL}{Replication Index column: a name}

\item{LLOQ}{LLOQ column: a name, an expression, or a fixed value}

\item{NBINS}{NBINS column: a name, an expression, or a fixed value}

\item{bin_by_strata}{Flag to indicate if the binning is by strata (default) or global}

\item{bin_style}{Style of the binning procedure: "ntile" (default), {"fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", or "jenks"} (ClassInterval styles), "breaks" (user defined breaks)}

\item{breaks}{User defined breaks: either a vector or a wide data.frame with a column "breaks", and other columns corresponding to the stratify variables}

\item{cut_right}{Flag to indicate if the intervals should be closed on the right (and open on the left) or vice versa. Default to False, unlike the cut function.}

\item{quantile_type}{Quantile algorithm: an integer between 4 and 9 selecting one of the nine quantile algorithms}

\item{filterblq}{Flag to remove the BLQ values just before computing the quantiles, but after "merging" with rep(, sim)}

\item{predcorrection}{Flag to indicate if the VPC has to be pred corrected or not}

\item{predcorrection_islogdv}{Flag to indicate if the data was log transformed}

\item{predcorrection_lowerbnd}{Not used yet}

\item{PI}{Prediction Intervals: vector of any number of  values}

\item{CI}{Confidence Intervals around the Prediction Intervals : vector of any number of value}

\item{bootstrapobsdata}{Not used yet}
}
\description{
Note: this function is deprecated in favor of the new "pipe-style" interface. See \code{\link{vpcstats}}.
}
\examples{

library(vpc)
library(vpcstats)
exampleobs <- simple_data$obs
exampleobs <- exampleobs[exampleobs$MDV == 0, ]
examplesim <- simple_data$sim
exampleobs$PRED <- examplesim[1:nrow(exampleobs), "PRED"]
exampleobs$REP <- 0
examplesim <- examplesim[examplesim$MDV == 0, ]#
exampleobs$LLOQ <- ifelse(exampleobs$ISM == 0, 100, 25)
examplesim$LLOQ <- ifelse(examplesim$ISM == 0, 100, 25)
VPCDATA<- vpcstats(
obsdata = exampleobs, simdata = examplesim, stratify = ~ISM,
NBINS = NULL, LLOQ = LLOQ)
}
