% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats2.R
\name{plot.vpcstatsobj}
\alias{plot.vpcstatsobj}
\title{Plot a \code{vpcstatsobj}.}
\usage{
\method{plot}{vpcstatsobj}(x, ..., show.points = TRUE,
  show.boundaries = TRUE, xlab = NULL, ylab = NULL,
  color = c("red", "blue", "red"), linetype = c("dotted", "solid",
  "dashed"), legend.position = "top")
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments can be specified but are ignored.}
}
\value{
A `ggplot` object.
}
\description{
Plot a \code{vpcstatsobj}.
}
